requirements.txt
appdirs==1.4.3
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
ciso8601==2.1.1
cssselect==1.0.3
fake-useragent==0.1.11
idna==2.7
lxml==4.2.5
mysql-connector==2.2.9
parse==1.9.0
pyee==5.0.0
pyppeteer==0.0.25
pyquery==1.4.0
requests==2.20.1
requests-html==0.9.0
selenium==3.141.0
six==1.11.0
tqdm==4.28.1
urllib3==1.24.1
w3lib==1.19.0
websockets==7.0
wget==3.2
